OrderingSystem/requirements.txt
Django >= 1.8.3
django-registration-redux >= 1.2.0
djangorestframework >= 3.1.3
geopy >= 1.10.0
#django-social-auth >= 0.7.20
python-social-auth >= 0.2.12
django-haystack